+++ preprocess.php
@@ -4,6 +4,11 @@
 	* @author Allan Thue Rehhoff
 	*/
 
+	// Cache headers
+	header("Cache-Control: no-store, no-cache, must-revalidate, max-age=0");
+	header("Cache-Control: post-check=0, pre-check=0", false);
+	header("Pragma: no-cache");
+
 	// Version check
 	if(version_compare(PHP_VERSION, 5.4, '<')) die("PHP >= 5.4 is required for this framework to function properly.");